renaming test resource package
